check if file exists before delete

+++ tests/Controllers/OriginalControllerTest.php
@@ -80,7 +80,9 @@ class OriginalControllerTest extends TestCase
         $fileName = 'balloons.jpg';
         $filePath = ROOT . '/tests/' . $fileName;
         $fs = new FileSystem();
-        unlink($fs->getStoragePath($fileName));
+        if (file_exists($fs->getStoragePath($fileName))) {
+            unlink($fs->getStoragePath($fileName));
+        }
         $fs->save([
             'name' => $fileName,
             'type' => 'image/jpeg',
